Move internal/k8sutil to internal/util/k8sutil
