+++ pkg/ansible/proxy/inject_owner.go
@@ -22,7 +22,7 @@ import (
 	"net/http"
 	"net/http/httputil"
 
-	"github.com/operator-framework/operator-sdk/internal/k8sutil"
+	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
 	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/controllermap"
 	k8sRequest "github.com/operator-framework/operator-sdk/pkg/ansible/proxy/requestfactory"
 	"github.com/operator-framework/operator-sdk/pkg/handler"
